all_2

+++ part3/app/api/v1/reviews.py
@@ -1,4 +1,5 @@
 from flask_restx import Namespace, Resource, fields
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.services import facade
 
 api = Namespace('reviews', description='Review operations')
@@ -7,23 +8,32 @@ api = Namespace('reviews', description='Review operations')
 review_model = api.model('Review', {
     'text': fields.String(required=True, description='Text of the review'),
     'rating': fields.Integer(required=True, description='Rating of the place (1-5)'),
-    'user_id': fields.String(required=True, description='ID of the user'),
     'place_id': fields.String(required=True, description='ID of the place')
 })
 
 @api.route('/')
 class ReviewList(Resource):
+    @jwt_required()
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
     def post(self):
         """Register a new review"""
         # Placeholder for the logic to register a new review
+        current_user = get_jwt_identity()
         review_data = api.payload
-        new_review = facade.create_review(review_data)
-
-        if not new_review:
+        place = facade.get_place(review_data['place_id'])
+        if not place:
+            return {'error': 'Place not found'}, 404
+        if not review_data:
             return {'error': 'No review found'}, 400
+        if place.owner_id == current_user:
+            return {'error': 'You cannot review your own place'}, 400
+        if facade.user_already_reviewed_place(current_user, review_data['place_id']):
+            return {'error': 'You have already reviewed this place'}, 400
+        
+        review_data['user_id'] = current_user
+        new_review = facade.create_review(review_data)
         return {'id': new_review.id, 'text': new_review.text, 'rating': new_review.rating, 'user_id': new_review.user.id, 'place_id': new_review.place.id}, 201
 
     @api.response(200, 'List of reviews retrieved successfully')
@@ -36,14 +46,18 @@ class ReviewList(Resource):
 
 @api.route('/<review_id>')
 class ReviewResource(Resource):
+    @jwt_required()
     @api.response(200, 'Review details retrieved successfully')
     @api.response(404, 'Review not found')
     def get(self, review_id):
         """Get review details by ID"""
         # Placeholder for the logic to retrieve a review by ID
+        current_user = get_jwt_identity()
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
         return {'id': review.id, 'text': review.text, 'rating': review.rating, 'user_id': review.user.id, 'place_id': review.place.id}, 200
 
     @api.expect(review_model)
@@ -53,20 +67,28 @@ class ReviewResource(Resource):
     def put(self, review_id):
         """Update a review's information"""
         # Placeholder for the logic to update a review by ID
+        current_user = get_jwt_identity()
         review_data = api.payload
-        review = facade.update_review(review_id, review_data)
-
+        review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 400
+        if review.user.id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+        updated_review = facade.update_review(review_id, review_data)
         return {'message': 'Review updated successfully'}, 200
-        
+
+    @jwt_required()    
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
     def delete(self, review_id):
         """Delete a review"""
         # Placeholder for the logic to delete a review
-        deleted_review = facade.delete_review(review_id)
-
-        if not deleted_review:
+        current_user = get_jwt_identity()
+        review = facade.get_review(review_id)
+        if not review:
             return {'error': 'Review not found'}, 404
+        if review.user_id != current_user:
+            return {'error': 'Unauthorized action'}, 403
+
+        deleted_review = facade.delete_review(review_id)
         return {'message': 'Review deleted successfully'}, 200
